v1.1.0harmattan50-free

+++ Q3E/src/main/jni/tdm/glslprogs/stages/interaction/interaction.ambient.fs.glsl
@@ -1,3 +1,4 @@
+#version 320 es
 /*****************************************************************************
 The Dark Mod GPL Source Code
 
@@ -12,9 +13,8 @@ or (at your option) any later version. For details, see LICENSE.TXT.
 Project: The Dark Mod (http://www.thedarkmod.com/)
 
 ******************************************************************************/
-#version 320 es
 
-precision mediump float;
+precision highp float;
 
 #pragma tdm_include "tdm_lightproject.glsl"
 #pragma tdm_include "tdm_interaction.glsl"
